# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GROFO API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks, optionally filtered by date range
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      responses:
        "204":
          description: Deleted
...
  /videos:
    get:
      operationId: listVideos
      tags: [videos]
      summary: Browse the video catalog
      parameters:
        - name: category
          in: query
          schema:
            $ref: "#/components/schemas/VideoCategory"
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Videos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Video"
  /videos/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      operationId: getVideo
      tags: [videos]
      summary: Get a single video
      responses:
        "200":
          description: Video
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
...
  /playlists:
    get:
      operationId: listPlaylists
      tags: [playlists]
      summary: List the user's playlists
      responses:
        "200":
          description: Playlists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Playlist"
    post:
      operationId: createPlaylist
      tags: [playlists]
      summary: Create a playlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaylistInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Playlist"
  /playlists/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      operationId: getPlaylist
      tags: [playlists]
      summary: Get a playlist with its videos
      responses:
        "200":
          description: Playlist detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaylistDetail"
    delete:
      operationId: deletePlaylist
      tags: [playlists]
      summary: Delete a playlist
      responses:
        "204":
          description: Deleted
  /playlists/{id}/items:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    post:
      operationId: addPlaylistItem
      tags: [playlists]
      summary: Add a video to a playlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaylistItemInput"
      responses:
        "201":
          description: Playlist detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaylistDetail"
  /playlists/{id}/items/{videoId}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
      - name: videoId
        in: path
        required: true
        schema:
          type: integer
    delete:
      operationId: removePlaylistItem
      tags: [playlists]
      summary: Remove a video from a playlist
      responses:
        "204":
          description: Deleted
  /note-folders:
    get:
      operationId: listNoteFolders
      tags: [notes]
      summary: List note folders
      responses:
        "200":
          description: Note folders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NoteFolder"
    post:
      operationId: createNoteFolder
      tags: [notes]
      summary: Create a note folder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteFolderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NoteFolder"
  /note-folders/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    delete:
      operationId: deleteNoteFolder
      tags: [notes]
      summary: Delete a note folder (notes inside become unfiled)
      responses:
        "204":
          description: Deleted
  /notes:
    get:
      operationId: listNotes
      tags: [notes]
      summary: List notes, optionally filtered
      parameters:
        - name: folderId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"
    post:
      operationId: createNote
      tags: [notes]
      summary: Create a note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
  /notes/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      operationId: getNote
      tags: [notes]
      summary: Get a note
      responses:
        "200":
          description: Note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
    patch:
      operationId: updateNote
      tags: [notes]
      summary: Update a note (used for autosave)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteUpdate"
      responses:
        "200":
          description: Updated note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
    delete:
      operationId: deleteNote
      tags: [notes]
      summary: Delete a note
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    VideoCategory:
      type: string
      enum: [NEET, JEE, CLAT, UPSC, Languages]
    TaskPriority:
      type: string
      enum: [low, medium, high]
...
    Task:
      type: object
      required: [id, title, date, completed, priority, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        date:
          type: string
          format: date
        completed:
          type: boolean
        priority:
          $ref: "#/components/schemas/TaskPriority"
        createdAt:
          type: string
          format: date-time
    TaskInput:
      type: object
      required: [title, date]
      properties:
        title:
          type: string
          minLength: 1
        date:
          type: string
          format: date
        priority:
          $ref: "#/components/schemas/TaskPriority"
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        date:
          type: string
          format: date
        completed:
          type: boolean
        priority:
          $ref: "#/components/schemas/TaskPriority"
...
    Video:
      type: object
      required:
        - id
        - youtubeId
        - title
        - channel
        - category
        - durationSeconds
        - thumbnailUrl
      properties:
        id:
          type: integer
        youtubeId:
          type: string
        title:
          type: string
        channel:
          type: string
        category:
          $ref: "#/components/schemas/VideoCategory"
        durationSeconds:
          type: integer
        thumbnailUrl:
          type: string
...
    Playlist:
      type: object
      required: [id, name, itemCount, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        itemCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    PlaylistInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
    PlaylistItemInput:
      type: object
      required: [videoId]
      properties:
        videoId:
          type: integer
    PlaylistDetail:
      type: object
      required: [id, name, createdAt, videos]
      properties:
        id:
          type: integer
        name:
          type: string
        createdAt:
          type: string
          format: date-time
        videos:
          type: array
          items:
            $ref: "#/components/schemas/Video"
    NoteFolder:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        createdAt:
          type: string
          format: date-time
    NoteFolderInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
    Note:
      type: object
      required: [id, title, content, folderId, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        folderId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    NoteInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        content:
          type: string
        folderId:
          type: ["integer", "null"]
    NoteUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        content:
          type: string
        folderId:
          type: ["integer", "null"]
...